cleanup: merged all mockups into one MJS file + moved to lib/fixtures
