--- tests/fixtures/dataworker/mock-user-service-empty.mjs
@@ -1,5 +1,0 @@
-export class UserService {
-  static async getUsers() {
-    return [];
-  }
-}
