--- tests/fixtures/dataworker/mock-user-service-reject.mjs
@@ -1,5 +1,0 @@
-export class UserService {
-  static async getUsers() {
-    throw new Error("users fetch smoke failure");
-  }
-}
