--- tests/fixtures/dataworker/mock-validator-invalid.mjs
@@ -1,5 +1,0 @@
-export class NotifierValidator {
-  static validateConfiguration() {
-    return { result: false, info: "forced invalid configuration" };
-  }
-}
