--- tests/fixtures/dataworker/mock-validator-valid.mjs
@@ -1,5 +1,0 @@
-export class NotifierValidator {
-  static validateConfiguration() {
-    return { result: true, info: "forced valid configuration" };
-  }
-}
