--- tests/fixtures/dataworker/mock-wait-on-ok.mjs
@@ -1,4 +1,0 @@
-export default async function waitOn(options = {}) {
-  const resources = Array.isArray(options.resources) ? options.resources.join(",") : "";
-  console.log(`wait-on ok: ${resources}`);
-}
