--- tests/fixtures/dataworker/mock-wait-on-reject.mjs
@@ -1,3 +1,0 @@
-export default async function waitOn() {
-  throw new Error("wait-on smoke failure");
-}
