+++ tests/lib/fixtures/loader-waiton-reject.mjs
@@ -2,11 +2,14 @@ import path from "path";
 import { fileURLToPath, pathToFileURL } from "url";
 
 const FIXTURE_DIR = path.dirname(fileURLToPath(import.meta.url));
-const INVALID_VALIDATOR_URL = pathToFileURL(path.join(FIXTURE_DIR, "mock-validator-invalid.mjs")).href;
+const MOCKS_BASE_URL = pathToFileURL(path.join(FIXTURE_DIR, "data-worker-mocks.mjs")).href;
 
 export async function resolve(specifier, context, defaultResolve) {
   if (specifier === "../notifiers/core/NotifierValidator.js") {
-    return { url: INVALID_VALIDATOR_URL, shortCircuit: true };
+    return { url: `${MOCKS_BASE_URL}?validator=valid`, shortCircuit: true };
+  }
+  if (specifier === "wait-on") {
+    return { url: `${MOCKS_BASE_URL}?waitOn=reject`, shortCircuit: true };
   }
   return defaultResolve(specifier, context, defaultResolve);
 }
